refactor(WelcomeScreen): mover estilos para arquivo styles.js
