+++ mavis/src/WelcomeScreen/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import { StyleSheet } from 'react-native';
 import { Block, Text, Button } from 'galio-framework';
+import styles from './styles';
 
 export default function WelcomeScreen({ navigation }) {
   return (
@@ -24,26 +24,3 @@ export default function WelcomeScreen({ navigation }) {
     </Block>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    padding: 20,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#fff',
-  },
-  title: {
-    textAlign: 'center',
-    marginBottom: 20,
-    color: '#333',
-  },
-  description: {
-    textAlign: 'center',
-    marginBottom: 40,
-    color: '#555',
-  },
-  button: {
-    alignSelf: 'center',
-    width: '50%',
-  },
-});
