+++ mavis/src/WelcomeScreen/styles.js
@@ -1,0 +1,26 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    padding: 20,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#fff',
+  },
+  title: {
+    textAlign: 'center',
+    marginBottom: 20,
+    color: '#333',
+  },
+  description: {
+    textAlign: 'center',
+    marginBottom: 40,
+    color: '#555',
+  },
+  button: {
+    alignSelf: 'center',
+    width: '100%',
+  },
+});
+
+export default styles
